% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_bins.R
\name{cut_bins}
\alias{cut_bins}
\title{Bin continuous data into discrete intervals}
\usage{
cut_bins(x, bins, labels = NULL, include.lowest = TRUE)
}
\arguments{
\item{x}{Numeric vector to bin.}

\item{bins}{Number of bins or a vector of bin edges.}

\item{labels}{Labels for the bins (optional).}

\item{include.lowest}{Whether to include the lowest value in the first interval (default: TRUE).}
}
\value{
A factor indicating bin membership for each element of x.
}
\description{
Similar to pandas.cut, divides data into bins.
}
\examples{
cut_bins(1:10, bins=3)
}
