% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explode.R
\name{explode}
\alias{explode}
\title{Explode a list-column into long format}
\usage{
explode(df, col)
}
\arguments{
\item{df}{A data frame.}

\item{col}{The name of the list-column to explode (unquoted).}
}
\value{
A data frame with the list-column exploded into long format.
}
\description{
Turns each element of a list-column into a row, duplicating the other columns as needed.
}
\examples{
df <- data.frame(a = 1:2, b = I(list(1:2, 3:4)))
explode(df, b)
}
