% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_df.R
\name{query_df}
\alias{query_df}
\title{Query a data frame with a string expression}
\usage{
query_df(df, expr)
}
\arguments{
\item{df}{A data frame.}

\item{expr}{A string expression to evaluate in the context of the data frame.}
}
\value{
A filtered data frame.
}
\description{
Filter rows of a data frame using a string expression (like pandas.DataFrame.query).
}
\examples{
df <- data.frame(a = 1:5, b = 6:10)
query_df(df, 'a > 2 & b < 10')
}
